// lib/types/enterprise-register.ts
export type VnProvince = { code: number; name: string };
export type VnWard = { code: number; name: string };

export type FormDataState = {
  companyName: string;
  taxCode: string;
  businessFields: string[];
  provinceCode: string;
  wardCode: string;
  provinceName: string;
  wardName: string;
  addressDetail: string;
  website: string;
  representativeName: string;
  representativeTitle: string;
  phone: string;
  email: string;
};